requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.4
django-extensions==3.1.3
django-on-heroku==1.1.2
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.30.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-oauth==1.0.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
oauthlib==3.1.1
packaging==20.9
protobuf==3.17.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
